Reorganise test cases
